Added code to allow for rule groups in collapser REST payloads, to support UI templates like the current design. Added code to define ui groups for existing rules. Added initial support to fetch values for the general settings for active life duration and generation separation, and to save them from UI. Refactored some UI javascript for brevity, modularity. Added a new HTTP mapping for workspaceparams, but had to hack it into current installation. See also BPS-267.

+++ code/bps/services/workspace/src/main/java/edu/berkeley/bps/services/workspace/Person.java
@@ -41,7 +41,7 @@ public class Person extends Entity {
 	
 	public static final int DEFAULT_ACTIVE_LIFE_YRS = 15;
 	public static final long DEFAULT_GENERATION_OFFSET = 
-								TimeUtils.convertYearsToMillis(DEFAULT_ACTIVE_LIFE_YRS);
+								TimeUtils.convertYearsToMillis((double)DEFAULT_ACTIVE_LIFE_YRS);
 	public static final double DEFAULT_ACTIVE_LIFE_WINDOW = 
 		TimeUtils.getDefaultWindowForActiveLife(DEFAULT_ACTIVE_LIFE_YRS);
 	public static final double DEFAULT_ACTIVE_LIFE_STDDEV =
@@ -266,6 +266,7 @@ public class Person extends Entity {
 		return nQuals;
 	}
 
+	// TODO Need to make this more general, and respect the configuration for #qualifications
 	/**
 	 * @return true if this has a declared forename, and 1 and only one qualifier
 	 */
@@ -275,6 +276,7 @@ public class Person extends Entity {
 			((declaredFather!=null) != (declaredClan!=null));
 	}
 
+	// TODO Need to make this more general, and respect the configuration for #qualifications
 	/**
 	 * @return true if this has a declared forename, and at least two qualifiers
 	 */
